# openapi/mgmt-api.yaml
openapi: "3.0.3"
info:
  title: Management Service API
  description: ペット・カテゴリのマスタ管理を行うマイクロサービス
  version: "1.0.0"

servers:
  - url: http://localhost:8083/api/v1

paths:
  /pets:
    get:
      tags: [pet]
      summary: ペット一覧取得
      operationId: getPets
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [available, pending, sold]
        - name: categoryId
          in: query
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pet"

    post:
      tags: [pet]
      summary: ペット新規登録
      operationId: createPet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePetRequest"
      responses:
        "201":
          description: 作成成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pet"

  /pets/{petId}:
    get:
      tags: [pet]
      summary: ペット詳細取得
      operationId: getPetById
      parameters:
        - name: petId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pet"

    put:
      tags: [pet]
      summary: ペット更新
      operationId: updatePet
      parameters:
        - name: petId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePetRequest"
      responses:
        "200":
          description: 更新成功

    delete:
      tags: [pet]
      summary: ペット削除
      operationId: deletePet
      parameters:
        - name: petId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: 削除成功

  /categories:
    get:
      tags: [category]
      summary: カテゴリ一覧取得
      operationId: getCategories
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

    post:
      tags: [category]
      summary: カテゴリ作成
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: 作成成功

components:
  schemas:
    Pet:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        category:
          $ref: "#/components/schemas/Category"
        status:
          type: string
          enum: [available, pending, sold]
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time

    Category:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string

    CreatePetRequest:
      type: object
      required:
        - name
        - categoryId
      properties:
        name:
          type: string
          maxLength: 100
        categoryId:
          type: integer
          format: int64
        status:
          type: string
          enum: [available, pending, sold]
          default: available
        tags:
          type: array
          items:
            type: string

    UpdatePetRequest:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [available, pending, sold]

    CreateCategoryRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 50
        description:
          type: string